Use more threads for saving images
Previously we had 8 threads hardcoded to perform the saving of images.
For a set of 600 paths this took around 10 minutes on a 32 core machine.
Updating the thread count to match the number of cores our processor has
resulted in a 3.5 minute runtime for the same set of paths.

This change should also help performance on machines with less than
eight cores as we'll have less communication overhead and saving an
image consumes an entire CPU core anyways.