Add config option to set threads for saving images
In testing i've found that using around 1.5 times the CPU cores
available on a machine yields the best performance with saving images.

This is really important on machines with a lot of cores, we use 32 core
machines in our tests and using up all the cores cuts down wraith time
in half.

8 seems like a good default, this PR adds the ability to change the
value.