removed casperjs from the wraith gem dependencies
a) casperjs isn't necessarily a dependency. Some people only use PhantomJS
b) this is causing issues when we try to upgrade to phantomjs 2.x, which requires casperjs beta 1.1. By listing casperjs as a dependency inside the Wraith gemspec, we're installing the default version of casperjs available to Rubygems, and this becomes the casperjs executable even if we've installed casperjs through something like brew. Consequently, we're stuck with an outdated version of casperjs and that prevents us from moving to phantomjs 2.

We should trust developers to sort out their own phantom/casper dependencies/configurations.